Move ITWatcher to shardingsphere-integration-test-fixture module (#9805)
* Refactor ITWatcher

* For xml format

* Refactor ITWatcher

* Move ITWatcher to shardingsphere-integration-test-fixture module
