+++ shardingsphere-test/shardingsphere-integration-test/shardingsphere-integration-test-suite/src/test/java/org/apache/shardingsphere/test/integration/engine/it/SingleITCase.java
@@ -19,10 +19,9 @@ package org.apache.shardingsphere.test.integration.engine.it;
 
 import lombok.AccessLevel;
 import lombok.Getter;
-import lombok.SneakyThrows;
 import org.apache.shardingsphere.test.integration.cases.dataset.DataSet;
 import org.apache.shardingsphere.test.integration.cases.dataset.DataSetLoader;
-import org.apache.shardingsphere.test.integration.engine.junit.ITWatcher;
+import org.apache.shardingsphere.test.integration.junit.watcher.ITWatcher;
 import org.junit.Before;
 import org.junit.Rule;
 
@@ -35,7 +34,6 @@ public abstract class SingleITCase extends BaseITCase {
     private DataSet dataSet;
     
     @Before
-    @SneakyThrows
     public void setup() {
         dataSet = null == getAssertion() ? null : DataSetLoader.load(getParentPath(), getDescription().getScenario(), getDescription().getDatabaseType(), getAssertion().getExpectedDataFile());
     }
